# swagger.yaml
openapi: 3.0.0
info:
  title: SportsMatch API
  description: API para la gestión de clubs, canchas, usuarios, eventos y reservas en SportsMatch.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
paths:
...
  /clubauth:
    post:
      tags:
        - ClubAuth
      summary: Registrar un nuevo club
      description: Crea un nuevo club con los detalles proporcionados.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClubCreate'
      responses:
        '201':
          description: Club creado con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          description: Error en la validación de los datos de entrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: El correo electrónico ya está en uso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - ClubAuth
      summary: Iniciar sesión como club
      description: Autentica un club y devuelve un token de acceso.
      security:
        - BasicAuth: [ ]
      responses:
        '200':
          description: Inicio de sesión exitoso.
          headers:
            c-api-key:
              schema:
                type: string
                description: Token de acceso generado para el club.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubAuthResponse'
        '401':
          description: Credenciales incorrectas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Acceso denegado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Club:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
    ClubCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
        phoneNumber:
          type: string
...
    ClubAuthResponse:
      type: object
      properties:
        club:
          $ref: '#/components/schemas/Club'
        token:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        details:
          type: array
          items:
            type: string